Build Tokenizer object. Single utils.py (tokenizer.utils methods have been displaced as static method from the tokenizer class). Files have be rearranged and adapted (especially predict and predict_and_explain from pytorch_model) accordingly. Not tested yet.
